--- lib/app/features/feed/views/components/post/constants.dart
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: ice License 1.0
-
-class PostConstants {
-  PostConstants._();
-
-  static double minVerticalMediaAspectRatio = 0.85;
-
-  static double maxHorizontalMediaAspectRatio = 1.63;
-}
