+++ lib/app/features/feed/views/components/text_editor/components/custom_blocks/text_editor_single_image_block/text_editor_single_image_block.dart
@@ -3,6 +3,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_quill/flutter_quill.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:ion/app/extensions/extensions.dart';
 import 'package:ion/app/features/gallery/providers/gallery_provider.c.dart';
 import 'package:photo_manager_image_provider/photo_manager_image_provider.dart';
 
@@ -14,9 +15,7 @@ const textEditorSingleImageKey = 'text-editor-single-image';
 class TextEditorSingleImageEmbed extends CustomBlockEmbed {
   TextEditorSingleImageEmbed(String path) : super(textEditorSingleImageKey, path);
 
-  static BlockEmbed image(String path) {
-    return TextEditorSingleImageEmbed(path);
-  }
+  static BlockEmbed image(String path) => TextEditorSingleImageEmbed(path);
 }
 
 ///
@@ -42,22 +41,30 @@ class TextEditorSingleImageBuilder extends EmbedBuilder {
       mainAxisSize: MainAxisSize.min,
       crossAxisAlignment: CrossAxisAlignment.stretch,
       children: [
-        ClipRRect(
-          borderRadius: BorderRadius.circular(12),
-          child: Consumer(
-            builder: (context, ref, child) {
-              final assetEntity = ref.watch(assetEntityProvider(path)).valueOrNull;
-              if (assetEntity == null) {
-                return const SizedBox.shrink();
-              }
-              return Image(
-                image: AssetEntityImageProvider(
-                  assetEntity,
-                  isOriginal: false,
-                ),
-                fit: BoxFit.cover,
-              );
-            },
+        Padding(
+          padding: EdgeInsets.symmetric(vertical: 8.0.s),
+          child: ClipRRect(
+            borderRadius: BorderRadius.circular(12),
+            child: Consumer(
+              builder: (context, ref, child) {
+                final assetEntity = ref.watch(assetEntityProvider(path)).valueOrNull;
+                if (assetEntity == null) {
+                  return const SizedBox.shrink();
+                }
+                return ConstrainedBox(
+                  constraints: BoxConstraints(
+                    maxHeight: 385.0.s,
+                  ),
+                  child: Image(
+                    image: AssetEntityImageProvider(
+                      assetEntity,
+                      isOriginal: false,
+                    ),
+                    fit: BoxFit.cover,
+                  ),
+                );
+              },
+            ),
           ),
         ),
       ],
